completed signup page (at long last)!
this is pretty much the final version, as it’s (as far as we can tell)
fully functional, bootstrapped, and debugged. i’ll probably be adding
some comments down the line for clarity and revising the use of
menu.php to menu_sign_in.php if that’s what we end up going with. give
her a whirl and see if you can find any bugs/have any questions!
#totesexcited