master_schedule.php
Sorry, forgot about having someone create a master schedule for Erika (or other admins) to view.  This also adds an admin delete functionality for scheduled volunteers.  

Of note is that this file is built to be fully integrated into EAB's server and, as such, may not work entirely in all development servers immediately.  Just revise the pathnames as necessary.

Finally, this is the first file of an AJAX two-file system and requires master_schedule_table.php to work.